third_party/blink/web_tests/external/wpt/intersection-observer/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] IntersectionObserverEntry interface object length
  assert_equals: wrong value for IntersectionObserverEntry.length expected 1 but got 0
Harness: the test ran to completion.
